MediatR and Testing.

+++ SharpMUSH.Library/Models/DBref.cs
@@ -4,13 +4,13 @@
 	{
 		public DBRef(int number) => Number = number;
 
-		public DBRef(int number, int? milliseconds)
+		public DBRef(int number, long? milliseconds)
 		{
 			Number = number;
 			CreationMilliseconds = milliseconds;
 		}
 
 		public int Number { get; init; }
-		public int? CreationMilliseconds { get; init; }
+		public long? CreationMilliseconds { get; init; }
 	}
 }
